+++ src/day8/re_frame/trace/view/views.cljs
@@ -1,3 +1,0 @@
-(ns day8.re-frame.trace.view.views)
-
-(defn render [])