+++ src/day8/re_frame_10x/view/views.cljs
@@ -1,0 +1,3 @@
+(ns day8.re-frame-10x.view.views)
+
+(defn render [])